# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/sessions/{sessionId}/revoke:
    post:
      operationId: revokeAdminSession
      tags: [admin]
      summary: Revoke a selected browser session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revoked session
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/AdminTrackedSession"
  /admin/broadcasts:
    post:
      operationId: sendAdminBroadcast
      tags: [admin]
      summary: Send a live message to signed-in browsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "204":
          description: Sent
  /admin/accounts:
    get:
      operationId: listAdminAccounts
      tags: [admin]
      summary: Search and list accounts
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, active, blocked]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountList"
  /admin/accounts/{userId}:
    get:
      operationId: getAdminAccount
      tags: [admin]
      summary: Get an account's full detail for admin review
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountDetail"
    delete:
      operationId: deleteAdminAccount
      tags: [admin]
      summary: Permanently delete an account
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountActionInput"
      responses:
        "204":
          description: Deleted
  /admin/accounts/{userId}/block:
    post:
      operationId: blockAdminAccount
      tags: [admin]
      summary: Temporarily block an account and revoke its active sessions
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountActionInput"
      responses:
        "200":
          description: Updated account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountSummary"
  /admin/accounts/{userId}/unblock:
    post:
      operationId: unblockAdminAccount
      tags: [admin]
      summary: Lift a temporary block on an account
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAccountActionInput"
      responses:
        "200":
          description: Updated account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountSummary"
...
components:
  schemas:
...
    AdminTrackedSession:
      type: object
      required: [clientSessionId, userId, email, deviceLabel, createdAt, lastSeenAt, revokedAt, ipAddress, currentPage]
      properties:
        clientSessionId:
          type: string
        userId:
          type: string
        email:
          type: string
        deviceLabel:
          type: string
        createdAt:
          type: string
        lastSeenAt:
          type: string
        revokedAt:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
          description: Latest server-observed IP address for this browser session.
        currentPage:
          type: ["string", "null"]
          description: Latest client-reported route (path only), refreshed on the heartbeat cadence.
    BroadcastInput:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 70
        body:
          type: string
          minLength: 1
          maxLength: 320
...
    SelectedStudySubject:
      type: object
      required: [id, subjectId, customName, name, sortOrder, createdAt]
      properties:
        id: { type: string, format: uuid }
        subjectId: { type: ["string", "null"] }
        customName: { type: ["string", "null"] }
        name: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string }
...
    AdminAccountSummary:
      type: object
      required: [userId, email, status, createdAt, lastSignInAt, bannedUntil, sessionCount, lastSeenAt]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string }
        status: { type: string, enum: [active, blocked] }
        createdAt: { type: string }
        lastSignInAt: { type: ["string", "null"] }
        bannedUntil: { type: ["string", "null"] }
        sessionCount: { type: integer }
        lastSeenAt: { type: ["string", "null"] }
    AdminAccountList:
      type: object
      required: [accounts, total, page, limit]
      properties:
        accounts:
          type: array
          items: { $ref: "#/components/schemas/AdminAccountSummary" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    AdminAccountPreferences:
      type: object
      required: [educationLevel, studyProfile, learningGoals, updatedAt]
      properties:
        educationLevel: { type: string }
        studyProfile: { type: string }
        learningGoals:
          type: array
          items: { type: string }
        updatedAt: { type: string }
    AdminStudySpaceSummary:
      type: object
      required: [id, title, sourceType, status, level, createdAt, updatedAt, lastViewedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        sourceType: { type: string, enum: [catalog, ai] }
        status: { type: string, enum: [draft, active] }
        level: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        lastViewedAt: { type: ["string", "null"] }
    AdminAccountAction:
      type: object
      required: [id, actorId, action, reason, result, errorDetail, createdAt]
      properties:
        id: { type: string, format: uuid }
        actorId: { type: string, format: uuid }
        action: { type: string, enum: [block, unblock, delete] }
        reason: { type: string }
        result: { type: string, enum: [success, failed] }
        errorDetail: { type: ["string", "null"] }
        createdAt: { type: string }
    AdminAccountDetail:
      type: object
      required: [summary, selectedSubjects, preferences, studySpaces, recentSessions, recentActions]
      properties:
        summary: { $ref: "#/components/schemas/AdminAccountSummary" }
        selectedSubjects:
          type: array
          items: { $ref: "#/components/schemas/SelectedStudySubject" }
        preferences:
          anyOf:
            - $ref: "#/components/schemas/AdminAccountPreferences"
            - type: "null"
        studySpaces:
          type: array
          items: { $ref: "#/components/schemas/AdminStudySpaceSummary" }
        recentSessions:
          type: array
          items: { $ref: "#/components/schemas/AdminTrackedSession" }
        recentActions:
          type: array
          items: { $ref: "#/components/schemas/AdminAccountAction" }
    AdminAccountActionInput:
      type: object
      required: [reason, idempotencyKey]
      properties:
        reason: { type: string, minLength: 1, maxLength: 500 }
        idempotencyKey: { type: string, minLength: 1, maxLength: 160 }