# lib/api-zod/src/generated/types/adminAccountDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAccountAction } from './adminAccountAction';
import type { AdminAccountPreferences } from './adminAccountPreferences';
import type { AdminAccountSummary } from './adminAccountSummary';
import type { AdminStudySpaceSummary } from './adminStudySpaceSummary';
import type { AdminTrackedSession } from './adminTrackedSession';
import type { SelectedStudySubject } from './selectedStudySubject';

export interface AdminAccountDetail {
  summary: AdminAccountSummary;
  selectedSubjects: SelectedStudySubject[];
  preferences: AdminAccountPreferences | null;
  studySpaces: AdminStudySpaceSummary[];
  recentSessions: AdminTrackedSession[];
  recentActions: AdminAccountAction[];
}