# lib/api-zod/src/generated/types/adminAccountPreferences.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAccountPreferences {
  educationLevel: string;
  studyProfile: string;
  learningGoals: string[];
  updatedAt: string;
}