# lib/api-zod/src/generated/types/adminTrackedSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminTrackedSession {
  clientSessionId: string;
  userId: string;
  email: string;
  deviceLabel: string;
  createdAt: string;
  lastSeenAt: string;
  /** @nullable */
  revokedAt: string | null;
  /**
     * Latest server-observed IP address for this browser session.
     * @nullable
     */
  ipAddress: string | null;
  /**
     * Latest client-reported route (path only), refreshed on the heartbeat cadence.
     * @nullable
     */
  currentPage: string | null;
}