# lib/api-zod/src/generated/types/selectedStudySubject.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SelectedStudySubject {
  id: string;
  /** @nullable */
  subjectId: string | null;
  /** @nullable */
  customName: string | null;
  name: string;
  sortOrder: number;
  createdAt: string;
}